Move the logic of the fiberExtractoin back into the pipeline task. Do cleanup.
